fix container bus register
