+++ src/main/java/net/minecraftforge/fml/common/CleanroomContainer.java
@@ -1,5 +1,8 @@
 package net.minecraftforge.fml.common;
 
+import com.google.common.eventbus.EventBus;
+import net.minecraftforge.common.ForgeVersion;
+
 public class CleanroomContainer {
     public static class Cleanroom extends DummyModContainer{
         public Cleanroom() {
@@ -8,7 +11,7 @@ public class CleanroomContainer {
             meta.modId = "cleanroom";
             meta.name = "Cleanroom";
             meta.description = "Cleanroom Minecraft.";
-            meta.version = ForgeVersion.forge;
+            meta.version = ForgeVersion.getVersion();
             meta.authorList.add("CleanroomMC");
         }
 
@@ -18,18 +21,38 @@ public class CleanroomContainer {
         }
     }
 
-    public MixinContainer() {
-        super(new ModMetadata());
-        ModMetadata meta = this.getMetadata();
-        meta.modId = "mixinbooter";
-        meta.name = "MixinBooter";
-        meta.description = "A Mixin library and loader.";
-        meta.version = "10.0";
-        meta.authorList.add("Rongmario");
+    @SuppressWarnings("unused")
+    public static class MixinContainer extends DummyModContainer{
+        public MixinContainer() {
+            super(new ModMetadata());
+            ModMetadata meta = this.getMetadata();
+            meta.modId = "mixinbooter";
+            meta.name = "MixinBooter";
+            meta.description = "A Mixin library and loader.";
+            meta.version = "10.0";
+            meta.authorList.add("Rongmario");
+        }
+
+        @Override
+        public boolean registerBus(EventBus bus, LoadController controller) {
+            return true;
+        }
     }
 
-    @Override
-    public boolean registerBus(EventBus bus, LoadController controller) {
-        return true;
+    public static class ConfigAnytimeContainer extends DummyModContainer{
+        public ConfigAnytimeContainer() {
+            super(new ModMetadata());
+            ModMetadata meta = this.getMetadata();
+            meta.modId = "configanytime";
+            meta.name = "ConfigAnytime";
+            meta.description = "Allows Forge configurations to be setup at any point in time.";
+            meta.version = "2.0";
+            meta.authorList.add("Rongmario");
+        }
+
+        @Override
+        public boolean registerBus(EventBus bus, LoadController controller) {
+            return true;
+        }
     }
 }
